Removed parameter in arena_alloc()
It's removed temporarily because I implemented that change without thinking about the rest of the code. I have to think of a capacity as a hint, rather than trusting the input from the user. That's because we reserve memory pages in multiples of 4096 (usually, not always) and that requires careful thinking before implementing this change