Add DynamicArray with C++ templates
It's experimental, because the interface is probably going to change in the future. I don't even know if I'm going to ditch the templates entirely if favor of C macros and I'm not sure if I should implement the functions as C++ methods directly.

Further investigation is needed to figure out if there are better ways to implement generic dynamic arrays in C style. Or maybe generic arrays in C are just stupid and I shouldn't implement it at all. Time will tell.